Refactor shell_command_helper to a lib folder
