+++ python_lib/__init__.py
@@ -1,0 +1,3 @@
+"""Lib module for daq sub units"""
+
+pass
